+++ src/background/windows_api/mod.rs
@@ -1,6 +1,7 @@
 mod app_bar;
 mod com;
 pub mod event_window;
+pub mod hdc;
 mod iterator;
 pub mod monitor;
 pub mod process;